Day 1: Solve day 1.
Just bloody lay out the notes in order and get elves to agree on what each one says.

Attempt to call locations in question. With phone. Even if not all of them have anyone else to pick up a phone, it'd drastically cut down on search time.

Receive word that the Chief Historian is present in a location. In fifty locations, even!

Receive explanation that elf phones can call sideways in time. Because I wished that the CH is where I called, it connected me to a timeline in which he is there.