Day 3: Solve day 3
Attempt to educate elf on the concept of prefix-free code, as evidently any instruction that ends in do() or don't() is equivalent to just writing do() or don't().

Choose not to mention the use of offsite code repositories for literally this situation.

Covertly slip their corrupted memory onto a storage device. It can't not be useful later, right?