Day 8: Solve day 8
Attempt to explain to elves that none of this is how any of this works.

Safely deconstruct mind control emitter technology and stash in bag for later I mean for safe disposal.

Finally get that we're running through a bunch of references to older AoCs because we're chasing the chief historian.

Feel a little concerned that this may be the last AoC.