Day 10: Solve day 10
Attempt to educate deer on my disappointment at this challenge's lack of complexity, but fail to make a compelling point because it is only easy due to years of experience.

Attempt to educate deer on the simplicity of pathfinding algorithms with a little thought, but fail to realise that deer does not speak English.

Attempt to improve CreedAssassinator, but fail to iterate upon its performance-costing parts.