Day 12: Solve day 12
Attempt to educate elves on that literally putting a fence around every 1x1 square would be the cheapest option, because multiplication.

Consider with dread that mathematicians have indubitably found an elegant way to determine the side count of arbitrary shapes in a discrete coordinate system.

To my credit, I also developed a rather elegant way to find the total number of sides within an enclosed area of plots. It's just useless, so it won't be written on this margin.