Day 13: Solve day 13.
Attempt to educate self on that every claw machine has a trivial solution that makes it seem like a nail.

Choose not to consider that in order for the buttons to move the claw by even n picometres each press - 1% of the size of a carbon atom, which is in itself 1 millionth the width of a human hair - in order for the prize to be 10 trillion picometers away, the machines would each have to be quite large, about 10 meters each.

Feel a modest amount of indignation at being made to do literal middle school math.

Decide that the indignation is simply within the maths being the solution unto itself; The problem makes it sound as though there were multiple solutions for each machine, as with buttons being multiples of each other, but this never happens. Each machine has exactly one solution. So instead of an optimisation problem, we have a trivial linear equation system solution.

This isn't even maths, this is baiting typos and accidentally calling the wrong variable. As a certain devil put it, if you're calculating a specific number, like with specific numbers, it ain't real maths.

Get distracted wondering whether the A and B buttons are coloured blue and brown.