Day 21: Solve day 21
Attempt to educate elves on the concept of a ship's computer that is in any way competent.

Choose not to point out that we are trying to open a door from where the historian is trapped, to depressurized space. Any issues with that maybe?

Attempt to educate elves on that they all have phones. Canonically. They work on hyperlight frequencies to allow communication across the solar system. I have one of them.