Day 22: Solve day 22
Attempt to educate elves on the concept of gun.

Decide not to further optimise challenge 2 because that would go against the spirit of trading BananaCoin.

Feel a little sad I didn't get to make a new pet this year.

Don't worry, the turtles are back at home with a MOBA player taking care of them. I can count on them to feed regularly!

Hope no work colleagues involved in inviting a certain CEO - if deserving of that title - read this day's solver class name.

Decide against making any crypto trader related jokes. Someone with that kind of portfolio doesn't need to be kicked further.

What's that rustling in bag?