Day 23: Speed up day 23
There's almost certainly a solution here that doesn't involve re-instantiating lists.

Since we're looking for a single largest network, we could track the largest network's size and return a single List that has the largest size instead of all possible Lists. This is left as an exercise to the reader :3

Yes, spanTheTree could be used for the first challenge too, but it's slow :< I know I usually would do just that, but I'm on a timer here and I'm so close to doing the entire thing in 2 seconds with 90% fitting in 1 second.